# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a Catopia API of CPE Senior Project.",
        "title": "Catopia API",
        "termsOfService": "http://somewhere.com/",
        "contact": {
            "name": "CPE34 - Catopia"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Login and get token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "operationId": "LoginHandler",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "User",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.Login"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/payload.LoginResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Register new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "operationId": "RegisterHandler",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "User",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.Regis"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/cat": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all cat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cat"
                ],
                "summary": "All Cat",
                "operationId": "GetAllCatHandler",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Cat"
                            }
                        }
                    }
                }
            }
        },
        "/api/cat/": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create new cat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cat"
                ],
                "operationId": "CreateCatHandler",
                "parameters": [
                    {
                        "description": "cat",
                        "name": "cat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.CreateCat"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/cat/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete cat by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cat"
                ],
                "operationId": "DeleteCatHandler",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of cat",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update cat by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cat"
                ],
                "operationId": "UpdateCatHandler",
                "parameters": [
                    {
                        "description": "cat",
                        "name": "cat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.UpdateCat"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "id of cat",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/recommend/cat": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get cat by user data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recommend"
                ],
                "operationId": "RecommendGetByUserHandler",
                "responses": {}
            }
        },
        "/api/recommend/cat/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get cat by cat ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recommend"
                ],
                "operationId": "RecommendGetByCatIDHandler",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of cat",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "domain.Cat": {
            "type": "object",
            "properties": {
                "aggression": {
                    "type": "integer"
                },
                "breeding": {
                    "type": "string"
                },
                "createAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "extraversion": {
                    "type": "integer"
                },
                "gender": {
                    "type": "string"
                },
                "group_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "last_update": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "profile": {
                    "type": "string"
                },
                "shyness": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
...
        "payload.CreateCat": {
            "type": "object",
            "required": [
                "breeding",
                "date",
                "gender",
                "name",
                "weight"
            ],
            "properties": {
                "aggression": {
                    "type": "integer",
                    "example": 5
                },
                "breeding": {
                    "type": "string",
                    "example": "siamese"
                },
                "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2021-01-20"
                },
                "extraversion": {
                    "type": "integer",
                    "example": 5
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female"
                    ],
                    "example": "male"
                },
                "name": {
                    "type": "string",
                    "example": "mori"
                },
                "profile": {
                    "type": "string",
                    "example": "url of image"
                },
                "shyness": {
                    "type": "integer",
                    "example": 5
                },
                "weight": {
                    "type": "number",
                    "format": "float",
                    "example": 3.5
                }
            }
        },
...
        "payload.Login": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "type": "string",
                    "example": "kongsakchai"
                }
            }
        },
        "payload.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "payload.Regis": {
            "type": "object",
            "required": [
                "date",
                "email",
                "gender",
                "password",
                "username"
            ],
            "properties": {
                "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2021-01-20"
                },
                "email": {
                    "type": "string",
                    "example": "email@email.com"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female"
                    ],
                    "example": "male"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "type": "string",
                    "example": "kongsakchai"
                }
            }
        },
        "payload.UpdateCat": {
            "type": "object",
            "properties": {
                "aggression": {
                    "type": "integer",
                    "example": 5
                },
                "breeding": {
                    "type": "string",
                    "example": "siamese"
                },
                "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2021-01-20"
                },
                "extraversion": {
                    "type": "integer",
                    "example": 5
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female"
                    ],
                    "example": "male"
                },
                "name": {
                    "type": "string",
                    "example": "mori"
                },
                "profile": {
                    "type": "string",
                    "example": "url of image"
                },
                "shyness": {
                    "type": "integer",
                    "example": 5
                },
                "weight": {
                    "type": "number",
                    "format": "float",
                    "example": 3.5
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}